Removed QT-Config editor for steamOS and LinuxOS
Reasoning: Linux and SteamOS for some reason yuzu is storing data in a very unreasonable way making it impossible or at the very least VERY hard for me to make the function work, these changes also improve the internal logic for Windows. A small patch but will fix issues for Linux again.